Refactoring - Outsource DPT 9.xxx
